==> step9.ipynb <==
Prasanna Hegde  
422985

Here, we combine [dataset_B]  and the [dataset_A : Train]

import pandas as pd

df_B = pd.read_csv('df_B.csv', encoding='utf-8')

df_B.head()

df_A = pd.read_csv("selected_pairs1.csv", encoding='utf-8')

from sklearn.model_selection import train_test_split

train_A, test_A = train_test_split(df_A, test_size=0.2, random_state=42)
print(f"items in training and test test are: {len(train_A)} and  {len(test_A)}")

df_C = pd.concat([train_A, df_B], ignore_index=True)  # concat dataset B and train_A

df_C.shape


df_C = df_C.sample(frac=1, random_state=42).reset_index(drop=True) #frac means fraction: 1 implies complete dataset is shuffled

df_C.head()

df_C.to_csv('df_C.csv', encoding='utf-8', index=False)
